reimplement toggle for privileged eSIM management
Use it only for controlling the LPA app state, EuiccSupportPixel package is now handled separately.
