+++ res/values/strings_ext.xml
@@ -24,4 +24,7 @@
     <string name="scramble_pin_title">PIN scrambling</string>
     <string name="scramble_pin_summary">Controls PIN scrambling option when inputting PIN on screen lock.</string>
 
+    <string name="privileged_euicc_management_title">Enable privileged eSIM management</string>
+    <string name="privileged_euicc_management_summary">Requires sandboxed Google Play installation</string>
+
 </resources>
